:test_tube: Test: Test파일 구조 변경
- 수정 파일
.\traffic\tests.py
.\traffic\models.py

- 수정 사항
tests.py 분리
models.py price 추가
